targeted-cleanup-with-some-failures
