+++ tests/modeling/tpc_ds_duckdb/aggregates/zquery_timing_amd64-intel64_family_6_model_183_stepping_1_genuineintel-16.log
@@ -1,4 +1,4 @@
 [query_03]
-parse_time = 0.051166
-exec_time = 0.029231
-comp_time = 0.111085
+parse_time = 0.053916
+exec_time = 0.038427
+comp_time = 1.509342
