+++ tests/scripts/config_directory/raw/users_with_pk.preql
@@ -1,5 +1,5 @@
 # Datasource ingested from users_with_pk
-# Generated on 2026-05-01 13:47:54.082305
+# Generated on 2026-05-02 10:00:17.699529
 import std.net;
 
 key user_id int;
