+++ tests/scripts/config_directory/raw/world_capitals.preql
@@ -1,5 +1,5 @@
 # Datasource ingested from world_capitals
-# Generated on 2026-05-01 13:47:56.056273
+# Generated on 2026-05-02 10:00:20.841847
 import std.geography;
 
 key country string::country;
