+++ trilogy/core/processing/discovery_utility.py
@@ -14,7 +14,6 @@ from trilogy.core.models.build import (
     BuildConcept,
     BuildConditional,
     BuildDatasource,
-    BuildFilterItem,
     BuildFunction,
     BuildGrain,
     BuildRowsetItem,
@@ -27,6 +26,7 @@ from trilogy.core.processing.condition_utility import (
     decompose_condition,
 )
 from trilogy.core.processing.constants import ROOT_DERIVATIONS
+from trilogy.core.processing.grain_utility import concept_source_address
 from trilogy.core.processing.nodes import GroupNode, MergeNode, StrategyNode
 from trilogy.core.processing.utility import GroupRequiredResponse
 from trilogy.utility import unique
@@ -191,15 +191,7 @@ def check_if_group_required(
     ngrain = []
     for con in target_grain.components:
         full = environment.concepts[con]
-        if full.derivation == Derivation.ROWSET:
-            assert isinstance(full.lineage, BuildRowsetItem)
-            ngrain.append(full.lineage.content.address)
-        elif full.derivation == Derivation.FILTER:
-            assert isinstance(full.lineage, BuildFilterItem)
-            if isinstance(full.lineage.content, BuildConcept):
-                ngrain.append(full.lineage.content.address)
-        else:
-            ngrain.append(full.address)
+        ngrain.append(concept_source_address(full))
     target_grain2 = BuildGrain.from_concepts(
         ngrain,
         environment=environment,
